fix(react-di): remove lodash dependency
BREAKING CHANGE: useDependencyInjection allows nested array no more

Do it manually or provide abstraction on useDependencyInjection.
This fixes esm target. The way lodash was included was not handled well in case of esm target.
Also, this makes it so it is not bundled for those who do not need it.